backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Hugo Thomas", "Tristan Le Saux","Manh-Huan Nguyen","Julien Perrier","Roberto Chacon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pandas = "^2.2.2"
pdfplumber = "^0.11.0"
nltk = "^3.8.1"
rdflib = "^7.0.0"
sparqlwrapper = "^2.0.0"
networkx = "2.6.3"
pytest = "^8.2.1"
pattern = "3.6"
deep-translator = "^1.11.4"
wikipedia-api = "^0.6.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
python-decouple = "^3.8"
pytest-asyncio = "^0.23.7"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.black]
line-length = 88

[tool.ruff]
select = ["E", "F", "I"]
fixable = ["ALL"]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true
plugins = "pydantic.mypy"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
